+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFWKZ6HNTJN84N95ASP43AP7"
+        "01KFWV1XHVS47C378FKQ7SATWB"
     )
     r"""The unique identifier of the data source"""