+++ packages/orq-rc/src/orq_ai_sdk/models/createidentityop.py
@@ -137,7 +137,7 @@ class CreateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-26T07:40:39.795Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-26T09:44:29.020Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")