+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -186,7 +186,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kfwkz5zpqee7j0hpqqwygxfs"
+        "conv_01kfwv1x05evfzsjq0pg78pxg1"
     )
     r"""Unique conversation identifier with `conv_` prefix."""