+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFWKZ6HRYK0H76TN4K3R64GD"
+        "01KFWV1XHZRMB5W79W83F31AH4"
     )
     r"""The unique identifier of the data source"""