+++ packages/orq-rc/RELEASES.md
@@ -5039,3 +5039,13 @@ Based on:
 - [python v4.3.0-rc.1] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.1 - packages/orq-rc
+
+## 2026-01-26 09:46:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.697.5 (2.799.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.2] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.2 - packages/orq-rc